+++ tests/phone_number_tests.rs
@@ -29,6 +29,6 @@ async fn verify_code_works() -> Result<(), WhatsappError> {
 }
 
 fn setup() {
-    dotenv::dotenv().ok();
+    dotenvy::dotenv().ok();
     let _ = env_logger::builder().is_test(true).try_init();
 }