+++ tests/send_message_tests.rs
@@ -122,6 +122,6 @@ async fn send_image_link_message_works() -> Result<(), WhatsappError> {
 }
 
 fn setup() {
-    dotenv::dotenv().ok();
+    dotenvy::dotenv().ok();
     let _ = env_logger::builder().is_test(true).try_init();
 }